Replace pipenv with poetry
Due to persistent issues with pipenv, with the most recent one being
that the lock file takes so long to get created, switching to another
dependency manager seemed fitting.

Ended up choosing poetry cause of it's simplicity and almost similar
style to pipenv.

Also, added support for python3.7 in the travis-Ci configuration.

Signed-off-by: Victor Otieno Omondi <vickz84259@gmail.com>